requirements/base.txt
Django
djangorestframework
django-cors-headers
django-filter
psycopg
channels
channels-redis
Pillow
celery
redis
django-treebeard
python-dotenv